fix(apistatusmessage): moved component to molecules
