+++ source/components/molecules/ApiStatusMessage/ApiStatusMessage.styled.ts
@@ -1,6 +1,6 @@
 import styled from "styled-components/native";
 
-import type { ThemeType } from "../../styles/themeHelpers";
+import type { ThemeType } from "../../../styles/themeHelpers";
 
 interface ApiStatusMessageContainerProps {
   theme: ThemeType;
