+++ source/components/molecules/ApiStatusMessage/ApiStatusMessage.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 
-import { Icon, Text } from "../../components/atoms";
+import { Icon, Text } from "../../atoms";
 
 import ApiStatusMessageContainer from "./ApiStatusMessage.styled";
-import theme from "../../styles/theme";
+import theme from "../../../styles/theme";
 
 import type { Props } from "./ApiStatusMessage.types";
