+++ source/components/molecules/index.ts
@@ -20,3 +20,4 @@ export { default as FloatingButton } from "./FloatingButton";
 export { default as DateTimeCard } from "./DateTimeCard";
 export { default as AddressCard } from "./AddressCard";
 export { default as BackNavigation } from "./BackNavigation";
+export { default as ApiStatusMessage } from "./ApiStatusMessage/ApiStatusMessage";
